server/requirements.txt
openenv[core]>=0.2.0
fastapi>=0.115.0
uvicorn>=0.24.0
pillow>=10.0.0
